backend/pyproject.toml
[tool.poetry]
name = "fastapi-async-sqlmodel-boilerplate"
version = "0.1.0"
description = "An opinionated yet highly extensible and fully asynchronous FastAPI boilerplate, embracing the Django-like folder structure. 🚀 Built with FastAPI, SQLModel (Pydantic v2 + SQLAlchemy 2.0), Alembic, PostgreSQL, Redis and ARQ. 📈"
authors = ["joaoflaviosantos <joaoflavio.s@hotmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
alembic = "^1.13.1"
arq = "^0.25.0"
asyncpg = "^0.29.0"
bcrypt = "^4.1.2"
fastapi = {extras = ["all"], version = "^0.109.0"}
fastapi-async-sqlalchemy = "^0.6.1"
fastapi-pagination = {extras = ["sqlalchemy"], version = "^0.12.14"}
gunicorn = "^21.2.0"
pytest = "^7.4.4"
pydantic-settings = "^2.1.0"
pyjwt = { extras = ["crypto"], version = "^2.8.0" }
python = "^3.11"
python-jose = "^3.3.0"
redis = "^5.0.1"
types-redis = "^4.6.0.11"
SQLAlchemy = "^2.0.23"
SQLAlchemy-Utils = "^0.41.1"
sqlmodel = "^0.0.14"
uvicorn = "^0.27.0"
uvloop = { version = "^0.19.0", markers = "sys_platform == 'linux'" }
psutil = "^5.9.8"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.6.0"
black = "^24.1.0"
mypy = "^1.8.0"

[tool.black]
line-length = 100

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
